backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
httpx>=0.28.0
openai>=1.60.0
websockets>=14.0
python-multipart>=0.0.18
numpy>=2.0.0
croniter>=2.0.0
